Java / Jakarta 21 updates
