--- Dockerfile
@@ -1,2 +1,0 @@
-FROM airhacks/glassfish
-COPY ./target/Java2XML.war ${DEPLOYMENT_DIR}
