+++ src/main/java/de/tgl/Java2XML/JAXRSConfiguration.java
@@ -5,6 +5,5 @@ import jakarta.ws.rs.core.Application;
 
 @ApplicationPath("/api")
 public class JAXRSConfiguration extends Application {
-    // No implementation needed
 }
