+++ src/main/java/de/tgl/Java2XML/models/Simple.java
@@ -1,4 +1,4 @@
-package de.tgl.Java2XML.obj;
+package de.tgl.Java2XML.models;
 
 public record Simple(int x, int y) {
 }
