--- src/main/webapp/index.html
@@ -1,3 +1,0 @@
-<html>
-hello world
-</html>
